requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
jinja2==3.1.6
apscheduler==3.11.0
python-dateutil==2.9.0.post0
pytest==8.4.1
